backend/requirements.txt
tensorflow==2.10.0
numpy==1.23.5
django==5.2.12
djangorestframework==3.17.1
django-cors-headers==4.4.0
opencv-python==4.13.0.92
pillow==12.2.0